backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.7  # PostgreSQL sync driver (backup)

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Caching & Background Tasks
redis>=5.0.1
celery>=5.3.4

# Data Processing & Analytics
pandas>=2.1.3
numpy>=1.25.2
requests>=2.31.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Production & Monitoring
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Stripe for payments
stripe>=7.7.0

# CORS and middleware
slowapi>=0.1.9  # Rate limiting
python-cors>=1.7.0 